+++ backend-lib/src/main/rust/lib.rs
@@ -381,6 +381,22 @@ fn encode_usk<'a>(
     )
 }
 
+fn encode_transaction<'a>(
+    env: &mut JNIEnv<'a>,
+    height: u64,
+    txid_bytes: Vec<u8>,
+) -> jni::errors::Result<JObject<'a>> {
+    let java_byte_array = env.byte_array_from_slice(&txid_bytes)?;
+    env.new_object(
+        "cash/z/ecc/android/sdk/internal/model/JniTransaction",
+        "(J[B)V",
+        &[
+            JValue::Long(height as jlong),
+            (&java_byte_array).into(),
+        ],
+    )
+}
+
 fn decode_usk(env: &JNIEnv, usk: JByteArray) -> anyhow::Result<UnifiedSpendingKey> {
     let usk_bytes = secret_from_jni(env, usk)?;
 
@@ -2848,7 +2864,7 @@ pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorClient_connectTo
 
 /// Frees a lightwalletd connection.
 #[unsafe(no_mangle)]
-pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_freeLightwalletdConnection<
+pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorWalletClient_freeLightwalletdConnection<
     'local,
 >(
     _: JNIEnv<'local>,
@@ -2864,7 +2880,7 @@ pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_freeLigh
 
 /// Returns information about this lightwalletd instance and the blockchain.
 #[unsafe(no_mangle)]
-pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_getServerInfo<'local>(
+pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorWalletClient_getServerInfo<'local>(
     mut env: JNIEnv<'local>,
     _: JClass<'local>,
     lwd_conn: jlong,
@@ -2883,7 +2899,7 @@ pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_getServe
 
 /// Returns information about this lightwalletd instance and the blockchain.
 #[unsafe(no_mangle)]
-pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_getLatestBlock<'local>(
+pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorWalletClient_getLatestBlock<'local>(
     mut env: JNIEnv<'local>,
     _: JClass<'local>,
     lwd_conn: jlong,
@@ -2902,12 +2918,12 @@ pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_getLates
 
 /// Fetches the transaction with the given ID.
 #[unsafe(no_mangle)]
-pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_fetchTransaction<'local>(
+pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorWalletClient_fetchTransaction<'local>(
     mut env: JNIEnv<'local>,
     _: JClass<'local>,
     lwd_conn: jlong,
     txid_bytes: JByteArray<'local>,
-) -> jbyteArray {
+) -> jobject {
     let res = catch_unwind(&mut env, |env| {
         let lwd_conn = ptr::with_exposed_provenance_mut::<crate::tor::LwdConn>(lwd_conn as usize);
         let lwd_conn = unsafe { lwd_conn.as_mut() }
@@ -2917,16 +2933,16 @@ pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_fetchTra
         // we may as well confirm we were actually passed something shaped correctly.
         let txid = parse_txid(env, txid_bytes)?;
 
-        let (tx, _height) = lwd_conn.get_transaction(txid)?;
+        let (tx, height) = lwd_conn.get_transaction(txid)?;
 
-        Ok(utils::rust_bytes_to_java(env, &tx)?.into_raw())
+        Ok(encode_transaction(env, height, tx)?.into_raw())
     });
     unwrap_exc_or(&mut env, res, ptr::null_mut())
 }
 
 /// Submits a transaction to the Zcash network via the given lightwalletd connection.
 #[unsafe(no_mangle)]
-pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_submitTransaction<
+pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorWalletClient_submitTransaction<
     'local,
 >(
     mut env: JNIEnv<'local>,
@@ -2948,7 +2964,7 @@ pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_submitTr
 
 /// Fetches the note commitment tree state corresponding to the given block height.
 #[unsafe(no_mangle)]
-pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorLwdConn_getTreeState<'local>(
+pub extern "C" fn Java_cash_z_ecc_android_sdk_internal_model_TorWalletClient_getTreeState<'local>(
     mut env: JNIEnv<'local>,
     _: JClass<'local>,
     lwd_conn: jlong,