+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/SdkSynchronizer.kt
@@ -30,6 +30,7 @@ import cash.z.ecc.android.sdk.internal.ext.existsSuspend
 import cash.z.ecc.android.sdk.internal.ext.tryNull
 import cash.z.ecc.android.sdk.internal.jni.RustBackend
 import cash.z.ecc.android.sdk.internal.model.Checkpoint
+import cash.z.ecc.android.sdk.internal.model.TorClient
 import cash.z.ecc.android.sdk.internal.model.TreeState
 import cash.z.ecc.android.sdk.internal.model.ZcashProtocol
 import cash.z.ecc.android.sdk.internal.model.ext.toBlockHeight
@@ -50,7 +51,6 @@ import cash.z.ecc.android.sdk.model.AccountImportSetup
 import cash.z.ecc.android.sdk.model.AccountUsk
 import cash.z.ecc.android.sdk.model.AccountUuid
 import cash.z.ecc.android.sdk.model.BlockHeight
-import cash.z.ecc.android.sdk.model.FastestServersResult
 import cash.z.ecc.android.sdk.model.FetchFiatCurrencyResult
 import cash.z.ecc.android.sdk.model.FirstClassByteArray
 import cash.z.ecc.android.sdk.model.ObserveFiatCurrencyResult
@@ -75,10 +75,11 @@ import cash.z.ecc.android.sdk.type.AddressType.Transparent
 import cash.z.ecc.android.sdk.type.AddressType.Unified
 import cash.z.ecc.android.sdk.type.ConsensusMatchType
 import cash.z.ecc.android.sdk.type.ServerValidation
-import co.electriccoin.lightwallet.client.LightWalletClient
+import cash.z.ecc.android.sdk.util.WalletClientFactory
+import co.electriccoin.lightwallet.client.WalletClient
 import co.electriccoin.lightwallet.client.model.LightWalletEndpoint
 import co.electriccoin.lightwallet.client.model.Response
-import co.electriccoin.lightwallet.client.new
+import co.electriccoin.lightwallet.client.util.use
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -114,7 +115,6 @@ import java.io.File
 import java.util.Locale
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.coroutines.CoroutineContext
-import kotlin.time.Duration.Companion.seconds
 
 /**
  * A Synchronizer that attempts to remain operational, despite any number of errors that can occur.
@@ -141,6 +141,9 @@ class SdkSynchronizer private constructor(
     private val fetchFastestServers: FastestServerFetcher,
     private val fetchExchangeChangeUsd: UsdExchangeRateFetcher,
     private val preferenceProvider: PreferenceProvider,
+    private val torClient: TorClient,
+    private val walletClient: WalletClient,
+    private val walletClientFactory: WalletClientFactory,
 ) : CloseableSynchronizer {
     companion object {
         private sealed class InstanceState {
@@ -177,23 +180,27 @@ class SdkSynchronizer private constructor(
             fastestServerFetcher: FastestServerFetcher,
             fetchExchangeChangeUsd: UsdExchangeRateFetcher,
             preferenceProvider: PreferenceProvider,
+            torClient: TorClient,
+            walletClient: WalletClient,
+            walletClientFactory: WalletClientFactory
         ): CloseableSynchronizer {
             val synchronizerKey = SynchronizerKey(zcashNetwork, alias)
-
             return mutex.withLock {
                 waitForShutdown(synchronizerKey)
                 checkForExistingSynchronizers(synchronizerKey)
-
                 SdkSynchronizer(
-                    context,
-                    synchronizerKey,
-                    repository,
-                    txManager,
-                    processor,
-                    backend,
-                    fastestServerFetcher,
-                    fetchExchangeChangeUsd,
-                    preferenceProvider
+                    context = context,
+                    synchronizerKey = synchronizerKey,
+                    storage = repository,
+                    txManager = txManager,
+                    processor = processor,
+                    backend = backend,
+                    fetchFastestServers = fastestServerFetcher,
+                    fetchExchangeChangeUsd = fetchExchangeChangeUsd,
+                    preferenceProvider = preferenceProvider,
+                    torClient = torClient,
+                    walletClient = walletClient,
+                    walletClientFactory = walletClientFactory
                 ).apply {
                     instances[synchronizerKey] = InstanceState.Active
                     start()
@@ -400,10 +407,7 @@ class SdkSynchronizer private constructor(
     override val latestBirthdayHeight
         get() = processor.birthdayHeight
 
-    override suspend fun getFastestServers(
-        context: Context,
-        servers: List<LightWalletEndpoint>
-    ): Flow<FastestServersResult> = fetchFastestServers(context, servers)
+    override suspend fun getFastestServers(servers: List<LightWalletEndpoint>) = fetchFastestServers(servers)
 
     internal fun start() {
         coroutineScope.onReady()
@@ -417,6 +421,8 @@ class SdkSynchronizer private constructor(
             coroutineScope.launch {
                 Twig.info { "Stopping synchronizer $synchronizerKey…" }
                 processor.stop()
+                torClient.dispose()
+                walletClient.dispose()
                 fetchExchangeChangeUsd.dispose()
             }
 
@@ -441,6 +447,9 @@ class SdkSynchronizer private constructor(
                 coroutineScope.launch {
                     Twig.info { "Stopping synchronizer $synchronizerKey…" }
                     processor.stop()
+                    torClient.dispose()
+                    walletClient.dispose()
+                    fetchExchangeChangeUsd.dispose()
                 }
 
             instances[synchronizerKey] = InstanceState.ShuttingDown(shutdownJob)
@@ -589,6 +598,7 @@ class SdkSynchronizer private constructor(
                             onScanComplete()
                             SYNCED
                         }
+
                         is Stopped -> STOPPED
                         is Disconnected -> DISCONNECTED
                         is Syncing -> SYNCING
@@ -734,6 +744,7 @@ class SdkSynchronizer private constructor(
                     Twig.info { "Chain tip for recovery until param fetched: ${response.result.value}" }
                     runCatching { response.result.toBlockHeight() }.getOrNull()
                 }
+
                 is Response.Failure -> {
                     Twig.error {
                         "Chain tip fetch for recovery until failed with: ${response.toThrowable()}"
@@ -997,13 +1008,9 @@ class SdkSynchronizer private constructor(
         // Create a dedicated light wallet client for the validation
         // The single request timeout is changed from default to 5 seconds to speed up a possible custom server
         // endpoint validation
-
-        LightWalletClient
-            .new(
-                context = context,
-                lightWalletEndpoint = endpoint,
-                singleRequestTimeout = 5.seconds
-            ).use { lightWalletClient ->
+        walletClientFactory
+            .create(endpoint = endpoint)
+            .use { lightWalletClient ->
                 val remoteInfo =
                     when (val response = lightWalletClient.getServerInfo()) {
                         is Response.Success -> response.result
@@ -1144,11 +1151,6 @@ internal object DefaultSynchronizerFactory {
             backend
         )
 
-    fun defaultService(
-        context: Context,
-        lightWalletEndpoint: LightWalletEndpoint
-    ): LightWalletClient = LightWalletClient.new(context, lightWalletEndpoint)
-
     internal fun defaultEncoder(
         backend: TypesafeBackend,
         saplingParamTool: SaplingParamTool,
@@ -1156,18 +1158,14 @@ internal object DefaultSynchronizerFactory {
     ): TransactionEncoder = TransactionEncoderImpl(backend, saplingParamTool, repository)
 
     fun defaultDownloader(
-        service: LightWalletClient,
+        walletClient: WalletClient,
         blockStore: CompactBlockRepository
-    ): CompactBlockDownloader = CompactBlockDownloader(service, blockStore)
+    ): CompactBlockDownloader = CompactBlockDownloader(walletClient, blockStore)
 
     internal fun defaultTxManager(
         encoder: TransactionEncoder,
-        service: LightWalletClient
-    ): OutboundTransactionManager =
-        OutboundTransactionManagerImpl.new(
-            encoder,
-            service
-        )
+        service: WalletClient
+    ): OutboundTransactionManager = OutboundTransactionManagerImpl.new(encoder, service)
 
     internal fun defaultProcessor(
         backend: TypesafeBackend,